requirements.txt
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
Flask==3.0.0
Flask-Cors==4.0.0
google-api-core==2.14.0
google-auth==2.24.0
google-cloud-aiplatform==1.36.4
google-cloud-bigquery==3.13.0
google-cloud-core==2.3.3
google-cloud-resource-manager==1.10.4
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-oauth==1.0.1
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.12.7
grpcio==1.59.3
grpcio-status==1.59.3
idna==3.6
importlib-metadata==6.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.4
packaging==23.2
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
python-dateutil==2.8.2
requests==2.31.0
rsa==4.9
shapely==2.0.2
six==1.16.0
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
